Updating demos to work with TEMPLATED time steppes
