+++ src/problem/CMakeLists.txt
@@ -1,7 +1,7 @@
 # Add source files in order of dependence, the ones with no dependency
 # first then the others
 #SET(SRC_FILES ac_interpolator.cpp cc_newton_interpolator.cpp cc_newtons_method.tpl.cpp ac_problem.cpp ac_ivp_for_odes.cpp ac_mesh_free_problem.tpl.cpp)
-SET(SRC_FILES ac_problem.cpp ac_ibvp.cpp ac_ibvp_for_odes.cpp)
+SET(SRC_FILES ac_problem.cpp ac_ibvp.tpl.cpp)
 
 # Include current directory to the path
 #INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR})
